feat(nav): highlight current page in header
